+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Go API: Added ability to create and register `BatchBuffer` plugins.
+- New `system_window` buffer for processing message windows (sliding or tumbling) following the system clock.
 
 ## 3.52.0 - 2021-08-02